@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices;
 
-// ReSharper disable once CheckNamespace
-namespace System.Runtime.CompilerServices
-{
-    internal class IsExternalInit
-    {
-    }
-}
+internal class IsExternalInit { }
