@@ -1,12 +1,2 @@
-﻿using System;
-
-namespace MathCore.WebAPIClients.ConsoleTest
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
+﻿
+Console.WriteLine("End.");
